+++ CHANGELOG.md
@@ -4,10 +4,11 @@
 
 *Major*
 
-* Added the Seer role.
+* Added the Seer role. ([#4](https://github.com/hjwylde/werewolf/issues/4))
+* Removed the need to encode / decode to JSON for the state file. ([#9](https://github.com/hjwylde/werewolf/issues/9))
 
 #### v0.1.0.0
 
 *Major*
 
-* Initial implementation with Villagers and Werewolves.
+* Initial implementation with Villagers and Werewolves. ([#1](https://github.com/hjwylde/werewolf/issues/1))